Desktop/theta-main/requirements.txt
flask
google-genai
openai 
boto3>=1.28.0
ffmpeg-python
flask-login
flask-sqlalchemy
werkzeug
google-generativeai
python-dotenv
stripe
requests
Flask-Session
flask-pymongo
pymongo
fastapi
uvicorn
celery
python-multipart
sendgrid
flask-mail
imageio[ffmpeg]
rq
